// lib/api/generated/model/blogPostDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface BlogPostDto {
  id?: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  featuredImageUrl?: string | null;
  isPublished?: boolean;
  blogCategoryId?: number;
  /** @nullable */
  categoryName?: string | null;
  authorId?: number;
  /** @nullable */
  authorName?: string | null;
}