// lib/api/generated/model/createBlogPostCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface CreateBlogPostCommand {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  featuredImageUrl?: string | null;
  isPublished?: boolean;
  blogCategoryId?: number;
}