// lib/api/generated/model/updateBlogPostCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export interface UpdateBlogPostCommand {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  featuredImageUrl?: string | null;
  isPublished?: boolean;
  blogCategoryId?: number;
}